doc: update Python req. for sphinx (#1733)
* doc: update Python req. for sphinx

* Upgrade the sphinx/requirements.txt to use latest versions of the
  following modules:
  * requests==2.24.0
  * sphinx==3.2.1
  * sphinx-markdown==1.0.2
  * recommonmark==0.6.0
  * breathe==4.23.0
* Update the sphinx configuration to use recommonmark instead
  of pandoc_markdown.
* Tweak some of the .rst files according to some warnings emitted by
  sphinx.
  * TODO: Fix other warnings in another pull request.
* Tweak build-documentation.sh script to simply build the documentation
  but not upload it to github.io
* doc: update doc build flow
* Add documentation build job to github actions
* Update travis apt requirements
* Add cmake messages to help diagnosing when building documentation
  fails
* clean up build-documentation.sh script
  * remove dead code
  * terminate if installing python requirements fails
* clean up push-documentation.sh script
  * set user to "OPAE Bot", opae_robot@intel
* doc: update gh action for Python setup
* doc: cleanup gh action, build-documentation script

* remove the gh setup python action and rely on the distro version of
python
* because build-documentation.sh is installing python packages with
--user, add ~/.local/bin to PATH

* doc: add python setuptools pkg to gh actions job

* doc: fix some warning in sphinx build

* doc: add doc artifacts to github action

* docs: add link checker in ci

* doc: upload link-checker results.


* docs: use docker img for building docs

* Add Dockerfile that builds an image with required packages for
  building documentation
  * This file was used to create and upload a docker image
    opae/docs-builder
* Change CI to pull docker image and build the documentation with it

* docs: fix docker img name

* Update no-python-analysis.yml

+++ scripts/build-documentation.sh
@@ -1,27 +1,26 @@
 #!/bin/bash
-
-COMMIT_ID="$(git rev-parse HEAD)"
-TEMP=$((git describe --exact-match $COMMIT_ID) 2>&1)
-export PYTHONPATH=$PWD/lib/python2.7:$PWD/lib/python3.6
-
-if [[ $TEMP == *"fatal"* ]]; then
-    echo "No new tag detected, hence builiding latest docs from master"
-    git reset --hard origin/master
-fi
+set -x
+PYTHON_VERSION=$(python3 --version | cut -d ' ' -f2)
 
 #install requirements
-pip install --user -r doc/sphinx/requirements.txt
+python3 -m pip install --user -r doc/sphinx/requirements.txt
+if [ $? != 0 ];
+then
+	echo "failed to install Python requirements"
+	exit 1
+fi
 if [ -d mybuild_docs ];
 then
 	rm -rf mybuild_docs
 fi
-
 mkdir mybuild_docs
 pushd mybuild_docs
-
+export PYTHONPATH=$PWD/lib/python$PYTHON_VERSION
+export PATH=~/.local/bin:$PATH
 trap "popd" EXIT
 
-cmake .. -DOPAE_BUILD_SPHINX_DOC=ON
+cmake .. -DOPAE_BUILD_SPHINX_DOC=ON -DOPAE_PYTHON_VERSION=$PYTHON_VERSION
+make _opae -j $(nproc)
 make docs
 make manpages
 
@@ -31,14 +30,4 @@ then
 else
     echo "Spinx documentation build FAILED"
     exit 1
-fi
-
-
-
-if [[ $TEMP != *"fatal"* ]]; then
-    echo "New tag detected; Uploading documentation under new tag on opae.github.io"
-    /bin/bash ../scripts/push-documentation.sh new_tag
-else
-    echo "No new tag detected; Latest documentation will be uploaded to opae.github.io"
-    /bin/bash ../scripts/push-documentation.sh latest
 fi
